initial menu drawer implementation
This requires the android-menudrawer project[0] to be added in the same way
as ActionBarSherlock.  There is certainly still work to be done on the
menu drawer (particularly around UI), but this initial version should be
fully functional.  This commit doesn't touch the existing Dashboard
menu (so there is duplication); that will come in a subsequent change.
This also doesn't yet move the blog selector into the menu drawer.

[0]: http://simonvt.github.com/android-menudrawer/

git-svn-id: https://android.svn.wordpress.org/trunk@905 a6565c4b-a94d-4d28-bb26-aa9e5ab542c0